backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
openai>=1.54.0
python-dotenv>=1.0.1
pydantic>=2.10.0
python-multipart>=0.0.12
google-cloud-texttospeech>=2.16.0
requests>=2.31.0
imageio-ffmpeg>=0.4.9
elevenlabs>=1.0.0
yt-dlp>=2024.0.0
sse-starlette>=2.0.0